+++ docs/changelog.rst
@@ -7,6 +7,11 @@ Changelog
 v0.8.0 - *not yet released*
     - **REMOVED** ``gen-coveragerc`` and **deprecated** :func:`~igbpyutils.dev.generate_coveragerc`
         - Use https://pypi.org/project/coverage-simple-excludes/ instead.
+    - **Renamed** ``igbpyutils.dev.check_script_vs_lib`` to :func:`igbpyutils.dev.script_vs_lib.check_script_vs_lib`
+        - The former is still available but raises a deprecation warning.
+        - The ``known_shebangs`` argument may now be a regular expression.
+        - The ``exec_from_git`` option now uses the current index instead of the tree,
+          so uncommitted changes are correctly seen.
 
 v0.7.0 - Sun Oct  6 2024
     - Added :class:`~igbpyutils.error.LoggingStream` for better type checking